requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Script
Flask-Babel
Flask-Testing
Flask-Mail
Flask-Cache
Flask-Login
Flask-OpenID
geoalchemy
nose
mysql-python
requests